X device - fix compiler 'warning'
Bug #707502 "- -Wincompatible-pointer-types warning in devices/gdevxini.c"

This is probably an oversight from when we changed a load of variables
to size_t.

Seems odd that compilers (well gcc) should refuse to compile becuase of
a warning, but that's compilers. The pointer type is incorrect so let's
fix it.

+++ devices/gdevx.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2001-2023 Artifex Software, Inc.
+/* Copyright (C) 2001-2024 Artifex Software, Inc.
    All Rights Reserved.
 
    This software is provided AS-IS with no warranty, either express or
@@ -39,7 +39,7 @@ typedef struct gx_device_X_s {
     bool is_buffered;
     bool IsPageDevice;
     byte *buffer;		/* full-window image */
-    long buffer_size;
+    size_t buffer_size;
     gx_device_color_info        orig_color_info;
 
     /* An XImage object for writing bitmap images to the screen */
